% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsg_themr.R
\name{rsg_themr}
\alias{rsg_themr}
\title{Make plot RSG theme}
\usage{
rsg_themr(
  rsg_colors = c("#9D9999", "#E2D192", "#C55A11", "#3C485E", "#46566E", "#5C7292",
    "#F6BA92", "#F3EDD3", "#BCC6D5", "#595959", "#075861"),
  input_layout = "clean"
)
}
\description{
This function uses the ggthemr package to create RSG-style R plots. It includes
default colors from RSG.
}
